+++ packages/core/src/constants.ts
@@ -3,3 +3,6 @@ import { DefaultGuardType } from './types';
 export const STATE_DELIMITER = '.';
 export const DEFAULT_GUARD_TYPE: DefaultGuardType = 'xstate.guard';
 export const TARGETLESS_KEY = '';
+export const NULL_EVENT = '';
+export const STATE_IDENTIFIER = '#';
+export const WILDCARD = '*';
